--- git_repository.go
@@ -1,50 +1,0 @@
-package trdl
-
-import (
-	git "github.com/go-git/go-git/v5"
-	"github.com/go-git/go-git/v5/plumbing/transport/http"
-
-	trdlGit "github.com/werf/vault-plugin-secrets-trdl/pkg/git"
-)
-
-func cloneGitRepositoryBranch(url, gitBranch, username, password string) (*git.Repository, error) {
-	cloneGitOptions := trdlGit.CloneOptions{
-		BranchName:        gitBranch,
-		RecurseSubmodules: git.DefaultSubmoduleRecursionDepth,
-	}
-
-	if username != "" && password != "" {
-		cloneGitOptions.Auth = &http.BasicAuth{
-			Username: username,
-			Password: password,
-		}
-	}
-
-	gitRepo, err := trdlGit.CloneInMemory(url, cloneGitOptions)
-	if err != nil {
-		return nil, err
-	}
-
-	return gitRepo, nil
-}
-
-func cloneGitRepositoryTag(url, gitTag, username, password string) (*git.Repository, error) {
-	cloneGitOptions := trdlGit.CloneOptions{
-		TagName:           gitTag,
-		RecurseSubmodules: git.DefaultSubmoduleRecursionDepth,
-	}
-
-	if username != "" && password != "" {
-		cloneGitOptions.Auth = &http.BasicAuth{
-			Username: username,
-			Password: password,
-		}
-	}
-
-	gitRepo, err := trdlGit.CloneInMemory(url, cloneGitOptions)
-	if err != nil {
-		return nil, err
-	}
-
-	return gitRepo, nil
-}
